+++ spec/app/app/serializers/alba/user/empty_nested_serializer.rb
@@ -1,0 +1,6 @@
+module Alba
+  module User
+    class EmptyNestedSerializer < AuthorSerializer
+    end
+  end
+end