+++ spec/app/app/serializers/ams/user/empty_nested_serializer.rb
@@ -1,0 +1,6 @@
+module Ams
+  module User
+    class EmptyNestedSerializer < AuthorSerializer
+    end
+  end
+end